requirements.txt
streamlit
schedule
pandas
requests
python-dotenv